RD-1073 Move EditWidget to components directory
